Implemented inheritance for Player  w/ virt fn for betting, improved UX, refactored files for c++17
Virtual destructor and constructor on Player for the bet() to work in bot and user. (thanks to https://stackoverflow.com/questions/19466662/when-my-base-class-and-derived-have-the-same-function-does-it-have-to-be-virtua). Lower vending probability. Made poker.h use   std::vector<Player*> players and updated files accordingly.    Username & game round printing. Fixed empty deck draw. Finished most things for MVP besides the important bot EHS algorithm.